pyproject.toml
[project]
name = "memr-interns"
version = "0.1.0"
description = ""
authors = [
    {name = "tioluwa",email = "tioluwa.dev@gmail.com"}
]
license = {text = "mit"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "whitenoise (>=6.8.2,<7.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-phonenumber-field[phonenumbers] (>=8.0.0,<9.0.0)",
    "uuid7 (>=0.1.0,<0.2.0)",
    "django-otp (>=1.5.4,<2.0.0)",
    "rsa (>=4.9,<5.0) ; python_full_version == \"3.10.14\"",
    "django-timezone-field (>=7.1,<8.0) ; python_full_version == \"3.10.14\"",
    "dcrypt (>=0.0.6,<0.0.7) ; python_full_version == \"3.10.14\"",
    "daphne (>=4.1.2,<5.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "django-unfold (>=0.44.0,<0.45.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
